+++ wiretiger-console/src/main/java/org/hum/wiretiger/console/http/servlet/PipeDetailServlet.java
@@ -22,7 +22,7 @@ public class PipeDetailServlet extends HttpServlet {
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 		resp.setHeader("content-type", "text/html; charset=UTF-8");
-		resp.getWriter().print(JSON.toJSONString(pipeService.getById(Integer.parseInt(req.getParameter("id")))));
+		resp.getWriter().print(JSON.toJSONString(pipeService.getById(Long.parseLong(req.getParameter("id")))));
 		resp.getWriter().flush();
 		resp.getWriter().close();
 	}
